test(runner): split oversized test files ahead of new cases
runStagedFlows.test.ts and outerHop.install.test.ts both sat at the 250-line
cap. The runStagedFlows mocks move to a shared testUtils harness; the
carry-over cases move to their own file. No test changes.
